LUN-2829 Handle workflows that do not create the attribute unsaved_select_profiles.
This attribute is set with the profiles that are selected by the user.
The attribute is necessary because the many-to-one relationships cannot be created before the profile promo grid is not saved, so we cannot do it in the form save.
When the plugin is copied, this attribute will be missing, so that is the reason of the hasattr.

+++ cmsplugin_profile/models.py
@@ -88,11 +88,16 @@ class ProfilePromoGrid(CMSPlugin):
 
     def save(self, *args, **kwargs):
         ret_value = super(ProfilePromoGrid, self).save(*args, **kwargs)
+        self._save_selected_profiles()
+        return ret_value
+
+    def _save_selected_profiles(self):
+        if not hasattr(self, 'unsaved_selected_profiles'):
+            return
         self.selectedprofile_set.all().delete()
         for unsaved_selected_profile in self.unsaved_selected_profiles:
             unsaved_selected_profile.promo_grid = self
             unsaved_selected_profile.save()
-        return ret_value
 
 
 class SelectedProfile(models.Model):
